Splits up classes to be in their own files
