+++ join_map_lanes/build.gradle
@@ -2,7 +2,7 @@ apply from: '../project.gradle'
 
 description = 'Code corresponding to the Join Value Lanes cookbook'
 ext.moduleName = 'swim.basic'
-mainClassName='swim.basic.BasicPlane'
+mainClassName = 'swim.basic.BasicPlane'
 
 dependencies {
     api group: 'org.swimos', name: 'swim-api', version: version
