+++ join_map_lanes/src/main/java/swim/basic/StreetStatisticsAgent.java
@@ -1,0 +1,16 @@
+package swim.basic;
+
+import swim.api.SwimLane;
+import swim.api.agent.AbstractAgent;
+import swim.api.lane.MapLane;
+
+class StreetStatisticsAgent extends AbstractAgent {
+
+  /*
+      - Key: Street name
+      - Value: Street population
+   */
+  @SwimLane("state")
+  MapLane<String, Integer> streetStatistics = this.mapLane();
+
+}
